:recycle: refactor(admin): do not change name announcement -> notice
