+++ backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/domains/announcement/dto/response/AnnouncementResponseDto.java
@@ -1,4 +1,4 @@
-package com.depromeet.domains.notice.dto.response;
+package com.depromeet.domains.announcement.dto.response;
 
 
 import lombok.AllArgsConstructor;
