+++ backend/streetdrop-admin/streetdrop-admin-server/src/main/java/com/depromeet/domains/announcement/repository/AnnouncementRepository.java
@@ -1,4 +1,4 @@
-package com.depromeet.domains.notice.repository;
+package com.depromeet.domains.announcement.repository;
 
 import com.depromeet.announcement.Announcement;
 import org.springframework.data.domain.Page;
